(refactor) separated the files other payment mode from mpesa file
